+++ src/main/java/com/maigrand/overwatchdb/repository/AbilityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AbilityRepository extends JpaRepository<Ability, Long> {
+public interface AbilityRepository extends JpaRepository<Ability, Integer> {
 }
