+++ src/main/java/com/maigrand/overwatchdb/repository/HeroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HeroRepository extends JpaRepository<Hero, Long> {
+public interface HeroRepository extends JpaRepository<Hero, Integer> {
 }
