+++ src/main/java/com/maigrand/overwatchdb/repository/WeaponRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WeaponRepository extends JpaRepository<Weapon, Long> {
+public interface WeaponRepository extends JpaRepository<Weapon, Integer> {
 }
